task-1/requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
Naked==0.1.32
Pillow==9.5.0
pyasn1==0.5.0
pypng==0.20220715.0
PyQRCode==1.2.1
PyYAML==6.0
pyzbar==0.1.9
requests==2.30.0
rsa==4.9
shellescape==3.8.1
urllib3==2.0.2
